Refacto network + fix calculus
