--- to/connectors/detection_interface.cpp
@@ -1,266 +1,0 @@
-// This file detects the messages sent by the v2X Gatway, Parses the JSON
-// Message to obtain the relevant message type for futher processing
-
-// Created by: KCL
-
-// Modified by: Omar Nassef (KCL)
-
-#include "include/detection_interface.h"
-#include "logger.h"
-
-std::string incomplete_message = std::string();
-std::list<std::string> incomplete_messages;
-
-rapidjson::Document    parse(std::string readFromServer) {
-  rapidjson::Document document;
-  rapidjson::Document::AllocatorType& allocator = document.GetAllocator();
-  if(readFromServer == "RECONNECT"){
-    document.SetObject();
-    document.AddMember("type",rapidjson::Value() .SetString(std::string("RECONNECT").c_str(),allocator),allocator);
-  }
-  else if(readFromServer == "\n"){
-    document.SetObject();
-    document.AddMember("type",rapidjson::Value() .SetString(std::string("heart_beat").c_str(),allocator),allocator);
-  }
-  else{
-    document.Parse(readFromServer.c_str());
-  }
-  return document;
-}
-
-Detected_Road_User assignRoadUserVals(rapidjson::Document &document) {
-	Detected_Road_User values;
-  // if(!(document.IsObject())){
-  //   return values;
-  // }
-  //TODO: use int instead of double when possible
-  values.type = document["type"].GetString();
-  document.HasMember("context") ? values.context = document["context"].GetString() : values.context = "placeholder";
-  document.HasMember("origin") ? values.origin = document["origin"].GetString() : values.origin = "placeholder";
-  document.HasMember("version") ? values.version = document["version"].GetString() : values.version = "placeholder";
-  document.HasMember("timestamp") ? values.timestamp = document["timestamp"].GetUint64() : values.timestamp = 0;
-  document["message"].HasMember("uuid") ? values.uuid = document["message"]["uuid"].GetString() : values.uuid = "placeholder";
-  document["message"].HasMember("its_station_type") ? values.its_station_type = document["message"]["its_station_type"].GetString() : values.its_station_type = "placeholder";
-  document["message"].HasMember("connected") ? values.connected = document["message"]["connected"].GetBool() : values.connected = false;
-  document["message"]["position"].HasMember("latitude") ? values.latitude = document["message"]["position"]["latitude"].GetDouble() : values.latitude = 0.0;
-  document["message"]["position"].HasMember("longitude") ? values.longitude = document["message"]["position"]["longitude"].GetDouble() : values.longitude = 0.0;
-  document["message"].HasMember("position_type") ? values.position_type = document["message"]["position_type"].GetString() : values.position_type = "placeholder";
-  document["message"].HasMember("heading") ? values.heading = document["message"]["heading"].GetDouble() : values.heading = 0.0;
-  document["message"].HasMember("speed") ? values.speed = document["message"]["speed"].GetDouble() : values.speed = 0.0;
-  document["message"].HasMember("acceleration") ? values.acceleration = document["message"]["acceleration"].GetDouble() : values.acceleration = 0.0;
-  document["message"].HasMember("yaw_rate") ? values.yaw_rate = document["message"]["yaw_rate"].GetDouble() : values.yaw_rate = 0.0;
-  if (document["message"].HasMember("size")) {
-    document["message"]["size"].HasMember("length") ? values.length = document["message"]["size"]["length"].GetDouble() : values.length = 0.0;
-    document["message"]["size"].HasMember("width") ? values.width = document["message"]["size"]["width"].GetDouble() : values.width = 0.0;
-    document["message"]["size"].HasMember("height") ? values.height = document["message"]["size"]["height"].GetDouble() : values.height = 0.0;
-  }
-  document["message"].HasMember("existence_probability") ? values.existence_probability = document["message"]["existence_probability"].GetUint() : values.existence_probability = -1;
-  if (document["message"].HasMember("accuracy")) {
-    document["message"]["accuracy"].HasMember("position_semi_major_confidence") ? values.position_semi_major_confidence = document["message"]["accuracy"]["position_semi_major_confidence"].GetUint() : values.position_semi_major_confidence = -1;
-    document["message"]["accuracy"].HasMember("position_semi_minor_confidence") ? values.position_semi_minor_confidence = document["message"]["accuracy"]["position_semi_minor_confidence"].GetUint() : values.position_semi_minor_confidence = -1;
-    document["message"]["accuracy"].HasMember("position_semi_major_orientation") ? values.position_semi_major_orientation = document["message"]["accuracy"]["position_semi_major_orientation"].GetUint() : values.position_semi_major_orientation = -1;
-    document["message"]["accuracy"].HasMember("heading") ? values.heading_c = document["message"]["accuracy"]["heading"].GetUint() : values.heading_c = -1;
-    document["message"]["accuracy"].HasMember("speed") ? values.speed_c = document["message"]["accuracy"]["speed"].GetUint() : values.speed_c = -1;
-    document["message"]["accuracy"].HasMember("acceleration") ? values.acceleration_c = document["message"]["accuracy"]["acceleration"].GetUint() : values.acceleration_c = 0;
-    document["message"]["accuracy"].HasMember("yaw_rate") ? values.yaw_rate_c = document["message"]["accuracy"]["yaw_rate"].GetUint() : values.yaw_rate_c = 0;
-    if (document["message"]["accuracy"].HasMember("size")) {
-      document["message"]["accuracy"]["size"].HasMember("length") ? values.length_c = document["message"]["accuracy"]["size"]["length"].GetDouble() : values.length_c = 0.0;
-      document["message"]["accuracy"]["size"].HasMember("width") ? values.width_c = document["message"]["accuracy"]["size"]["width"].GetDouble() : values.width_c = 0.0;
-      document["message"]["accuracy"]["size"].HasMember("height") ? values.height_c = document["message"]["accuracy"]["size"]["height"].GetDouble() : values.height_c = 0.0;
-    }
-  }
-  document["message"].HasMember("color") ? values.color = document["message"]["color"].GetString() : values.color = "0x0000";
-  document["message"].HasMember("lane_position") ? values.lane_position = document["message"]["lane_position"].GetUint() : values.lane_position = 15;
-  document.HasMember("signature") ? values.signature = document["signature"].GetString() : values.signature = "placeholder";
-  document.HasMember("source_uuid") ? values.source_uuid = document["source_uuid"].GetString() : values.source_uuid = "placeholder";
-  document.HasMember("message_id") ? values.message_id = document["message_id"].GetString() : values.message_id = "placeholder";
-
-  return values;
-
-}
-
-std::vector<std::string> assignNotificationDeleteVals(rapidjson::Document    &document) {
-  std::vector<std::string> values;
-  if (document["message"].HasMember("ru_uuid_list") && document["message"]["ru_uuid_list"].IsArray()) {
-    const rapidjson::Value &array = document["message"]["ru_uuid_list"].GetArray();
-    for (rapidjson::SizeType i = 0; i < array.Size(); i++) {
-      values.emplace_back(array[i].GetString());
-    }
-  } else {
-    logger::write("error: notify_delete do not contain member ru_uuid_list.");
-  }
-  return values;
-}
-
-Detected_To_Notification assignNotificationVals(rapidjson::Document    &document) {
-  Detected_To_Notification values;
-  values.type = document["type"].GetString();
-  document.HasMember("context") ? values.context = document["context"].GetString() : values.context = "placeholder";
-  document.HasMember("origin") ? values.origin = document["origin"].GetString() : values.origin = "placeholder";
-  document.HasMember("version") ? values.version = document["version"].GetString() : values.version = "placeholder";
-  document.HasMember("timestamp") ? values.timestamp = document["timestamp"].GetUint64() : values.timestamp = -1;
-  document["message"].HasMember("subscription_id") ? values.subscriptionId = document["message"]["subscription_id"].GetInt() : values.subscriptionId = 0;
-  (document.HasMember("signature")) ? (values.signature = document["signature"].GetString()) : (values.signature = "placeholder");
-  (document.HasMember("source_uuid")) ? (values.source_uuid = document["source_uuid"].GetString()) : (values.source_uuid = "placeholder");
-  (document.HasMember("message_id")) ? (values.message_id = document["message_id"].GetString()) : (values.message_id = "placeholder");
-  logger::write("received notification with timestamp=" + std::to_string(values.timestamp));
-  for(auto& v : document["message"]["ru_description_list"].GetArray()) {
-    rapidjson::StringBuffer sb;
-    rapidjson::Writer<rapidjson::StringBuffer> writer(sb);
-    v.Accept(writer);
-    rapidjson::Document ruDocument = parse(sb.GetString());
-    const Detected_Road_User &ru = assignRoadUserVals(ruDocument);
-    if (ru.timestamp == 0){
-      logger::write("received invalid road user description in notify_add");
-    } else {
-      values.ru_description_list.push_back(ru);
-    }
-    sb.Clear();
-    writer.Reset(sb);
-  }
-
-
-  return values;
-}
-
-Detected_Trajectory_Feedback assignTrajectoryFeedbackVals(rapidjson::Document    &document) {
-  Detected_Trajectory_Feedback values;
-
-  document.HasMember("type") ? values.type = document["type"].GetString() : values.type = "placeholder";
-  document.HasMember("context") ? values.context = document["context"].GetString() : values.context = "placeholder";
-  document.HasMember("origin") ? values.origin = document["origin"].GetString() : values.origin = "placeholder";
-  document.HasMember("version") ? values.version = document["version"].GetString() : values.version = "placeholder";
-  document.HasMember("timestamp") ? values.timestamp = document["timestamp"].GetUint64() : values.timestamp = -1;
-  document.HasMember("source_uuid") ? values.uuid_vehicle = document["source_uuid"].GetString() : values.uuid_vehicle = "placeholder";
-  document.HasMember("destination_uuid") ? values.uuid_to = document["destination_uuid"].GetString() : values.uuid_to = "placeholder";
-  document["message"].HasMember("uuid_maneuver") ? values.uuid_maneuver = document["message"]["uuid_maneuver"].GetString() : values.uuid_maneuver = "placeholder";
-  document["message"].HasMember("timestamp") ? values.timestamp_message = document["message"]["timestamp"].GetUint64() : values.timestamp_message = -1;
-  document["message"].HasMember("feedback") ? values.feedback = document["message"]["feedback"].GetString() : values.feedback = "placeholder";
-  document["message"].HasMember("reason") ? values.reason = document["message"]["reason"].GetString() : values.reason = "placeholder";
-  (document.HasMember("signature")) ? (values.signature = document["signature"].GetString()) : (values.signature = "placeholder");
-  (document.HasMember("message_id")) ? (values.message_id = document["message_id"].GetString()) : (values.message_id = "placeholder");
-
-  return values;
-}
-
-
-Detected_Subscription_Response assignSubResponseVals(rapidjson::Document    &document) {
-
-  Detected_Subscription_Response values;
-
-  document.HasMember("type") ? values.type = document["type"].GetString() : values.type = "placeholder";
-  document.HasMember("context") ? values.context = document["context"].GetString() : values.context = "placeholder";
-  document.HasMember("origin") ? values.origin = document["origin"].GetString() : values.origin = "placeholder";
-  document.HasMember("version") ? values.version = document["version"].GetString() : values.version = "placeholder";
-  document.HasMember("timestamp") ? values.timestamp = document["timestamp"].GetUint64() : values.timestamp = -1;
-  document.HasMember("result") ? values.result = document["result"].GetString() : values.result = "placeholder";
-  document["message"].HasMember("request_id") ? values.request_id = document["message"]["request_id"].GetInt() : values.request_id = -1;
-  document["message"].HasMember("subscription_id") ? values.subscriptionId = document["message"]["subscription_id"].GetInt() : values.subscriptionId = -1;
-  (document.HasMember("signature")) ? (values.signature = document["signature"].GetString()) : (values.signature = "placeholder");
-  (document.HasMember("source_uuid")) ? (values.source_uuid = document["source_uuid"].GetString()) : (values.source_uuid = "placeholder");
-  (document.HasMember("message_id")) ? (values.message_id = document["message_id"].GetString()) : (values.message_id = "placeholder");
-
-  return values;
-
-}
-
-Detected_Unsubscription_Response assignUnsubResponseVals(rapidjson::Document    &document) {
-
-  Detected_Unsubscription_Response values;
-
-  document.HasMember("type") ? values.type = document["type"].GetString() : values.type = "placeholder";
-  document.HasMember("context") ? values.context = document["context"].GetString() : values.context = "placeholder";
-  document.HasMember("origin") ? values.origin = document["origin"].GetString() : values.origin = "placeholder";
-  document.HasMember("version") ? values.version = document["version"].GetString() : values.version = "placeholder";
-  document.HasMember("timestamp") ? values.timestamp = document["timestamp"].GetUint64() : values.timestamp = -1;
-  document["message"].HasMember("result") ? values.result = document["message"]["result"].GetInt() : values.result = -1;
-  document["message"].HasMember("request_id") ? values.request_id = document["message"]["request_id"].GetInt() : values.request_id = -1;
-  (document.HasMember("signature")) ? (values.signature = document["signature"].GetString()) : (values.signature = "placeholder");
-  (document.HasMember("source_uuid")) ? (values.source_uuid = document["source_uuid"].GetString()) : (values.source_uuid = "placeholder");
-  (document.HasMember("message_id")) ? (values.message_id = document["message_id"].GetString()) : (values.message_id = "placeholder");
-
-  return values;
-
-}
-
-message_type filterInput(rapidjson::Document    &document) {
-  if(!(document.IsObject())){
-    return message_type::unknown;
-  }
-  if(document["type"] == NOTIFY_ADD) {
-    return message_type::notify_add;
-  }
-
-  else if(document["type"] == RECONNECT) {
-    return message_type::reconnect;
-  }
-
-  else if(document["type"] == HEART_BEAT) {
-    return message_type::heart_beat;
-  }
-
-  else if(document["type"] == SUBSCRIPTION_RESPONSE) {
-    return message_type::subscription_response;
-  }
-
-  else if(document["type"] == UNSUBSCRIPTION_RESPONSE) {
-    return message_type::unsubscription_response;
-  }
-
-  else if(document["type"] == TRAJECTORY_FEEDBACK) {
-    return message_type::trajectory_feedback;
-  }
-
-  else if(document["type"] == NOTIFY_DELETE) {
-      return message_type::notify_delete;
-  }
-  else {
-    logger::write("error: received message with unknown type: " + std::string(document["type"].GetString()));
-    return message_type::unknown;
-  }
-
-}
-
-std::vector<std::string> listenDataTCP(int socket_c) {
-
-  char dataReceived[MAXIMUM_TRANSFER];
-  memset(dataReceived, 0, sizeof(dataReceived));
-  std::string to_return;
-  std::vector<std::string> returning;
-
-  while (returning.empty()) {
-    int i = read(socket_c, dataReceived, sizeof(dataReceived));
-
-    if(i < 0) {
-      logger::write("Error: Failed to receive transmitted data.\n");
-      logger::write("trying to reconnect.\n");
-      returning.push_back("RECONNECT");
-      return returning;
-    }
-    else if(i == 0) {
-      logger::write("Socket closed from the remote server.\n");
-      returning.push_back("RECONNECT");
-      return returning;
-    }
-    else if(i > 0){
-      for(int index = 0; index < i; index++){
-        char chrc = dataReceived[index];
-        if (chrc == '\n') {
-          to_return = incomplete_message;
-          incomplete_messages.push_back(to_return);
-          incomplete_message = std::string();
-        } else {
-          incomplete_message += chrc;
-        }
-      }
-    }
-
-    while(!incomplete_messages.empty()) {
-      auto message_packet = incomplete_messages.front();
-      if(message_packet != "\n" && message_packet != std::string()) logger::write(message_packet + "\n");
-      returning.emplace_back(incomplete_messages.front());
-      incomplete_messages.pop_front();
-    }
-  }
-  return returning;
-}
